requirements.txt
# --- Web Automation ---
playwright

# --- API Testing ---
requests

# --- Database Validation ---
mysql-connector-python

# --- Test Framework Runner---
pytest
pytest-html

# --- Utilities and config mngt---
pyyaml
python-dotenv

# --- Optional: If using rich console logging or JSON parsing ---
rich

# -- parallel execution
pytest-xdist